# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plog CMS API
  description: Plog CMS 系统 API 接口规范
  version: 1.0.0
  contact:
    name: Plog Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
  - url: https://admin.example.com/api/v1
    description: 生产环境
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Auth]
      summary: 用户登出
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /comments:
    get:
      tags: [Comments]
      summary: 获取评论列表
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentListResponse'
...
  /comments/{id}/approve:
    post:
      tags: [Comments]
      summary: 审核通过评论
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 审核成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    CookieAuth:
      type: apiKey
      in: cookie
      name: AUTH_COOKIE
    ApiKeyAuth:
      type: apiKey
      in: query
      name: api_key
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    User:
      type: object
      properties:
        uid:
          type: integer
        username:
          type: string
        nickname:
          type: string
        role:
          type: string
        email:
          type: string
        avatar:
          type: string
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            token:
              type: string
            user:
              $ref: '#/components/schemas/User'
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        gid:
          type: integer
        pid:
          type: integer
        content:
          type: string
        author:
          type: string
        email:
          type: string
        url:
          type: string
        ip:
          type: string
        date:
          type: string
        status:
          type: string
    CommentListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
        meta:
          type: object
          properties:
            request_id:
              type: string
              format: uuid
            timestamp:
              type: string
              format: date-time